fixup! net/traffic,ipn/ipnlocal: extract traffic steering utilities
Rename util/traffic to net/traffic
