+++ ipn/ipnlocal/local.go
@@ -62,6 +62,7 @@ import (
 	"tailscale.com/net/netns"
 	"tailscale.com/net/netutil"
 	"tailscale.com/net/packet"
+	"tailscale.com/net/traffic"
 	"tailscale.com/net/tsaddr"
 	"tailscale.com/net/tsdial"
 	"tailscale.com/paths"
@@ -95,7 +96,6 @@ import (
 	"tailscale.com/util/syspolicy/policyclient"
 	"tailscale.com/util/syspolicy/ptype"
 	"tailscale.com/util/testenv"
-	"tailscale.com/util/traffic"
 	"tailscale.com/util/usermetric"
 	"tailscale.com/util/vizerror"
 	"tailscale.com/version"
